+++ library/src/Auth.cpp
@@ -3,11 +3,13 @@
 #include <fcntl.h>
 #include <unistd.h>
 
+#include <crypto/Aes.hpp>
 #include <crypto/Random.hpp>
 #include <crypto/Sha256.hpp>
 
-#include <printer/Printer.hpp>
+#include <fs/Path.hpp>
 #include <fs/ViewFile.hpp>
+#include <printer/Printer.hpp>
 
 #include "sos/Auth.hpp"
 
@@ -23,8 +25,7 @@
 
 namespace printer {
 Printer &operator<<(Printer &printer, const sos::Auth::SignatureInfo &a) {
-  return printer
-    .key("size", var::NumberString(a.size()))
+  return printer.key("size", var::NumberString(a.size()))
     .key("hash", var::View(a.hash()).to_string<var::GeneralString>())
     .key("signature", a.signature().to_string());
 }
@@ -61,29 +62,29 @@ bool Auth::authenticate(var::View key) {
 
   auto random_token = start(token_out);
 
-  if( var::View(random_token.auth_token()).truncate(16) != var::View(random_data).truncate(16) ){
-    //first 16 bytes should stay the same
+  if (
+    var::View(random_token.auth_token()).truncate(16)
+    != var::View(random_data).truncate(16)) {
+    // first 16 bytes should stay the same
     return false;
   }
 
-  const auth_key_token_t hash_out_input = {
-    .key = token_key.auth_token(),
-    .token = random_token.auth_token()};
+  const auth_key_token_t hash_out_input
+    = {.key = token_key.auth_token(), .token = random_token.auth_token()};
 
-  const auth_key_token_t hash_in_input = {
-    .key = random_token.auth_token(),
-    .token = token_key.auth_token()};
+  const auth_key_token_t hash_in_input
+    = {.key = random_token.auth_token(), .token = token_key.auth_token()};
 
-  const auto hash_out = Token(
-    crypto::Sha256::get_hash(fs::ViewFile(var::View(hash_out_input))));
+  const auto hash_out
+    = Token(crypto::Sha256::get_hash(fs::ViewFile(var::View(hash_out_input))));
 
   // do SHA256 calcs
   const auto hash_in = finish(hash_out);
 
-  const auto hash_in_expected = Token(
-    crypto::Sha256::get_hash(fs::ViewFile(var::View(hash_in_input))));
+  const auto hash_in_expected
+    = Token(crypto::Sha256::get_hash(fs::ViewFile(var::View(hash_in_input))));
 
-  if( hash_in == hash_in_expected ){
+  if (hash_in == hash_in_expected) {
     return true;
   }
 
@@ -112,12 +113,9 @@ Auth::Token Auth::finish(const Token &token) {
   return Token(result);
 }
 
-Auth::SignatureInfo
-Auth::get_signature_info(const fs::FileObject &file) {
+Auth::SignatureInfo Auth::get_signature_info(const fs::FileObject &file) {
 
   fs::File::LocationScope ls(file);
-
-
 
   if (file.size() < sizeof(auth_signature_marker_t)) {
     return SignatureInfo();
@@ -125,12 +123,11 @@ Auth::get_signature_info(const fs::FileObject &file) {
 
   const auto signature = get_signature(file);
 
-  if( signature.is_valid() == false ){
+  if (signature.is_valid() == false) {
     return SignatureInfo();
   }
 
-  const size_t hash_size
-    = file.size() - sizeof(auth_signature_marker_t);
+  const size_t hash_size = file.size() - sizeof(auth_signature_marker_t);
 
   auto hash = [](const fs::FileObject &file, size_t hash_size) {
     crypto::Sha256 result;
@@ -140,19 +137,18 @@ Auth::get_signature_info(const fs::FileObject &file) {
     return result.output();
   }(file, hash_size);
 
-  return SignatureInfo().set_hash(hash).set_signature(signature).set_size(hash_size);
+  return SignatureInfo().set_hash(hash).set_signature(signature).set_size(
+    hash_size);
 }
 
-crypto::Dsa::Signature
-Auth::get_signature(const fs::FileObject &file) {
+crypto::Dsa::Signature Auth::get_signature(const fs::FileObject &file) {
   if (file.size() < sizeof(auth_signature_marker_t)) {
     return crypto::Dsa::Signature();
   }
 
   fs::File::LocationScope ls(file);
 
-  const size_t marker_location
-    = file.size() - sizeof(auth_signature_marker_t);
+  const size_t marker_location = file.size() - sizeof(auth_signature_marker_t);
   auth_signature_marker_t signature = {};
   file.seek(marker_location).read(var::View(signature));
 
@@ -166,7 +162,8 @@ Auth::get_signature(const fs::FileObject &file) {
   }
 }
 
-crypto::Dsa::Signature Auth::sign(const fs::FileObject & file, const crypto::Dsa & dsa){
+crypto::Dsa::Signature
+Auth::sign(const fs::FileObject &file, const crypto::Dsa &dsa) {
   fs::File::LocationScope ls(file);
   const auto hash = crypto::Sha256::get_hash(file.seek(0));
   const auto signature = dsa.sign(hash);
@@ -204,3 +201,121 @@ bool Auth::verify(
   return crypto::Dsa(crypto::Dsa::KeyPair().set_public_key(public_key))
     .verify(signature_info.signature(), signature_info.hash());
 }
+
+#if defined __link
+void Auth::create_secure_file(const CreateSecureFile &options) {
+  const var::PathString input_path = options.input_path();
+  const var::PathString secure_path = options.output_path();
+
+  const auto input_hash = crypto::Sha256::get_hash(fs::File(input_path));
+
+  fs::File padded_file
+    = fs::File(input_path, fs::OpenMode::append_write_only());
+  const u32 original_size = padded_file.size();
+  const u32 padding_required = 16 - (original_size % 16);
+
+  for (const auto i : api::Index(padding_required)) {
+    MCU_UNUSED_ARGUMENT(i);
+    char c = options.padding_character();
+    padded_file.write(var::View(c));
+  }
+
+  // hash and encrypt the file
+  crypto::Aes::Key encryption_key;
+
+  if (options.key().is_empty() == false) {
+    if (options.is_remove_key() == false) {
+      API_RETURN_ASSIGN_ERROR(
+        "you must create a secure archive when using a custom key",
+        EINVAL);
+    }
+    const auto option_key = crypto::Aes::Key::from_string(options.key());
+    encryption_key.set_key(option_key.get_key256());
+  }
+
+  const auto key_to_write = options.is_remove_key()
+                              ? crypto::Aes::Key().nullify().key256()
+                              : encryption_key.key256();
+
+  // writes the key and IV to the file
+  fs::File(fs::File::IsOverwrite::yes, secure_path)
+    .write(var::View(secure_file_version))
+    .write(var::View(original_size))
+    .write(key_to_write)
+    .write(encryption_key.initialization_vector())
+    .write(input_hash)
+    .write(
+      fs::File(input_path),
+      crypto::AesCbcEncrypter()
+        .set_initialization_vector(encryption_key.initialization_vector())
+        .set_key256(encryption_key.key256()),
+      fs::File::Write().set_progress_callback(options.progress_callback()));
+}
+
+void Auth::create_plain_file(const CreatePlainFile &options) {
+
+
+  fs::File source = fs::File(options.input_path());
+
+  u32 version = 0;
+  source.read(var::View(version));
+
+  u32 original_size = 0;
+
+  if (version == secure_file_version) {
+    source.read(var::View(original_size));
+  } else {
+    // older versions just started with original file size
+    original_size = version;
+  }
+
+  crypto::Aes::Key256 key_256;
+  crypto::Aes::InitializationVector iv;
+  crypto::Sha256::Hash input_hash;
+  source.read(key_256).read(iv);
+
+  if( version == secure_file_version){
+    source.read(input_hash);
+  }
+
+  if (source.is_error()) {
+    API_RETURN_ASSIGN_ERROR("failed to read metadata from source file", EINVAL);
+  }
+
+  if (options.key().is_empty() == false) {
+    key_256
+      = crypto::Aes::Key(crypto::Aes::Key::Construct().set_key(options.key()))
+          .key256();
+  } else {
+    if (crypto::Aes::Key(key_256).is_key_null()) {
+      // error
+      API_RETURN_ASSIGN_ERROR(
+        "no key was provided, but a key is required",
+        EINVAL);
+    }
+  }
+
+  fs::File output_file
+    = fs::File(fs::File::IsOverwrite::yes, options.output_path())
+        .write(
+          source,
+          crypto::AesCbcDecrypter().set_initialization_vector(iv).set_key256(
+            key_256),
+          fs::File::Write().set_progress_callback(options.progress_callback()))
+        .move();
+
+  // need File::truncate(path, size)
+  truncate(var::PathString(options.output_path()).cstring(), original_size);
+
+  if (version == secure_file_version) {
+    const auto check_input_hash
+      = crypto::Sha256::get_hash(fs::File(options.output_path()));
+    if (var::View(check_input_hash) != var::View(input_hash)) {
+      API_RETURN_ASSIGN_ERROR(
+        "failed to decrypt file and recover original Sha256 hash -- password "
+        "is probably wrong",
+        EINVAL);
+    }
+  }
+}
+#endif