+++ CHANGES.md
@@ -1,16 +1,24 @@
-# Version 1.1.0
+# Version 1.2.0
 
 ## New Features
 
 - Update `CMakeLists.txt` to require setup of SDK in a super-project
 - Add `Auth::create_secure_file` and `Auth::create_plain_file`
 - Add `set_orphan()` to `Appfs::FileAttributes`
-- Add `Sos::wait_pid()` method to access unistd `waitpid()`
 
 ## Bug Fixes
 
 - Fix an `EINVAL` error when using `Appfs::append`
 - Add `const` to `SerialNumber::operator==()`
+
+# Version 1.1.0
+
+## New Features
+
+- Add `Sos::wait_pid()` method to access unistd `waitpid()`
+
+## Bug Fixes
+
 - Minor cleanup for `Auth` class
 
 # Version 1.0.0